Fixed spacing issues between messages in Conversation Message List.
Also, CoversationScreenImpl is moved to the new file.
